+++ packages/axios-extension/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/axios-extension
 
+## 1.25.13
+
+### Patch Changes
+
+- Updated dependencies [dd2131c]
+    - @sap-ux/btp-utils@1.1.9
+
+## 1.25.12
+
+### Patch Changes
+
+- bda7356: fix(deps): update dependency qs to v6.14.2 [security]
+
 ## 1.25.11
 
 ### Patch Changes